requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anndata==0.12.10
annotated-types==0.7.0
archspec==0.2.5
array-api-compat==1.13.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
attrs==25.4.0
beautifulsoup4==4.14.3
biopython @ file:///croot/biopython_1755770124226/work
bottle==0.13.4
brotli==1.2.0
bs4==0.0.2
certifi==2026.1.4
charset-normalizer==3.4.4
cheroot==11.1.2
click==8.3.1
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
debugpy @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_debugpy_1769744980/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
donfig==0.8.1.post1
et_xmlfile==2.0.0
ete4==4.4.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastexcel==0.19.0
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gb-io==0.4.0
gecco==0.10.2
gecco-tool==0.10.0
genecat @ file:///home/benr/thesis/genecat/src
gitdb==4.0.12
GitPython==3.1.46
google-crc32c==1.8.0
h5py==3.15.1
idna==3.11
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipykernel_1770566133/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1770040907/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jaraco.functools==4.4.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_core_1760643864/work
kiwisolver==1.4.9
legacy-api-wrap==1.5
lightning-utilities==0.15.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
mkl-service==2.5.2
mkl_fft @ file:///home/task_176159249290440/conda-bld/mkl_fft_1761592897595/work
mkl_random @ file:///home/task_176159294306005/conda-bld/mkl_random_1761592958759/work
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
natsort==8.4.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.6.1
numcodecs==0.16.5
numpy @ file:///home/task_177073265815615/croot/numpy_and_numpy_base_1770732709806/work/dist/numpy-2.4.2-cp313-cp313-linux_x86_64.whl#sha256=9571aa986e3d2cc15b20222355ede05f68dc57d7554a0717493fe7f1f88c1a0e
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openpyxl==3.1.5
packaging @ file:///home/task_176104878999627/conda-bld/packaging_1761049078606/work
pandas==2.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1770676243/work
patsy==1.0.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pillow==12.1.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1771233827/work
polars==1.38.1
polars-runtime-32==1.38.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.4.1
protobuf==6.33.5
psutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_psutil_1769678153/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyhmmer==0.12.0
pyparsing==3.3.2
pyrodigal==3.7.0
pysqlite3==0.6.0
python-crfsuite==0.9.12
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyzmq_1771716961/work
requests==2.32.5
rich==14.3.3
rich-argparse==1.7.2
scikit-learn==1.8.0
scipy==1.17.1
sentry-sdk==2.53.0
setuptools==80.10.2
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sklearn-crfsuite==0.5.0
smmap==5.0.2
soupsieve==2.8.3
sqlite-vec==0.1.6
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels==0.14.6
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.10.0
torchmetrics==1.8.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1765836373784/work
tqdm==4.67.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
triton==3.6.0
typing-inspection==0.4.2
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1770634525285/work
wheel==0.46.3
yarl==1.22.0
zarr==3.1.5
